feat: layer(s) for switching windows, ability to add/remove current app from state

+++ install.sh
@@ -1,9 +1,13 @@
 set -e 
 
-ln -s "$PWD/scripts/iterm/itermctl.py" "$HOME/.local/bin/itermctl"
+if [ ! -e "$HOME/.local/bin/itermctl" ]; then
+    ln -s "$PWD/scripts/iterm/itermctl.py" "$HOME/.local/bin/itermctl"
+fi
 chmod +x "$HOME/.local/bin/itermctl"
 npm run build
 pip install iterm2
 
-ln -s "$PWD/scripts/keybindstate.py" "$HOME/.local/bin/keybindstate"
+if [ ! -e "$HOME/.local/bin/keybindstate" ]; then
+    ln -s "$PWD/scripts/keybindstate.py" "$HOME/.local/bin/keybindstate"
+fi
 chmod +x "$HOME/.local/bin/keybindstate"
